scheduler_api/pyproject.toml
[tool.poetry]
name = "scheduler-api"
version = "0.1.0"
description = ""


[tool.poetry.dependencies]
python = "^3.8"
fastapi = "*"
uvicorn = "*"
celery = "*"
python-dotenv = "*"
redis = "*"
psycopg = {extras = ["binary"], version = "*"}
sqlalchemy = "*"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"